annotated version of the code
contains annotated version of the code. Some redundant parts were removed, and I added some options for sorting the data. I think using both the upward and downward parts of the casting and sorting them by height yields the best results, but you can check it for yourself. I tried calculating the u-directional velocity shortly, but it did not work, so I removed it again.